fix: refactor bull queues to a common module for shared configuration
