+++ services/graph/apps/worker/src/graph_notifier/graph.monitor.processor.module.ts
@@ -1,14 +1,11 @@
 import { BlockchainModule, BlockchainService, ConfigModule, ConfigService, GraphStateManager } from '#lib';
-import { BullModule } from '@nestjs/bullmq';
 import { Module } from '@nestjs/common';
 import { RedisModule } from '@songkeys/nestjs-redis';
 import { GraphNotifierService } from './graph.monitor.processor.service';
-import * as QueueConstants from '#lib/utils/queues';
 
 @Module({
   imports: [
     BlockchainModule,
-    ConfigModule,
     RedisModule.forRootAsync(
       {
         imports: [ConfigModule],
@@ -19,55 +16,8 @@ import * as QueueConstants from '#lib/utils/queues';
       },
       true, // isGlobal
     ),
-    BullModule.forRootAsync({
-      imports: [ConfigModule],
-      useFactory: (configService: ConfigService) => {
-        // Note: BullMQ doesn't honor a URL for the Redis connection, and
-        // JS URL doesn't parse 'redis://' as a valid protocol, so we fool
-        // it by changing the URL to use 'http://' in order to parse out
-        // the host, port, username, password, etc.
-        // We could pass REDIS_HOST, REDIS_PORT, etc, in the environment, but
-        // trying to keep the # of environment variables from proliferating
-        const url = new URL(configService.redisUrl.toString().replace(/^redis[s]*/, 'http'));
-        const { hostname, port, username, password, pathname } = url;
-        return {
-          connection: {
-            host: hostname || undefined,
-            port: port ? Number(port) : undefined,
-            username: username || undefined,
-            password: password || undefined,
-            db: pathname?.length > 1 ? Number(pathname.slice(1)) : undefined,
-          },
-        };
-      },
-      inject: [ConfigService],
-    }),
-    BullModule.registerQueue(
-      {
-        name: QueueConstants.GRAPH_CHANGE_NOTIFY_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: true,
-          removeOnFail: false,
-          attempts: 3,
-          backoff: {
-            type: 'exponential',
-          },
-        },
-      },
-      {
-        name: QueueConstants.GRAPH_CHANGE_REQUEST_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: true,
-          removeOnFail: false,
-          attempts: 3,
-        },
-      },
-      {
-        name: QueueConstants.RECONNECT_REQUEST_QUEUE,
-      },
-    ),
   ],
-  providers: [GraphNotifierService, GraphStateManager, BlockchainService, ConfigService],
-  exports: [BullModule, GraphNotifierService, BlockchainService, ConfigService],
+  providers: [GraphNotifierService, GraphStateManager],
+  exports: [GraphNotifierService],
 })
 export class GraphNotifierModule {}
