+++ services/graph/apps/worker/src/graph_publisher/graph.publisher.processor.module.ts
@@ -1,69 +1,12 @@
-/*
-https://docs.nestjs.com/modules
-*/
-
-import { BullModule } from '@nestjs/bullmq';
 import { Module } from '@nestjs/common';
-import { RedisModule } from '@songkeys/nestjs-redis';
-import * as QueueConstants from '#lib/utils/queues';
-import { BlockchainModule, ConfigModule, ConfigService, NonceService } from '#lib';
+import { BlockchainModule, NonceService } from '#lib';
 import { GraphUpdatePublisherService } from './graph.publisher.processor.service';
 
 @Module({
   imports: [
     BlockchainModule,
-    ConfigModule,
-    RedisModule.forRootAsync(
-      {
-        imports: [ConfigModule],
-        useFactory: (configService: ConfigService) => ({
-          config: [{ url: configService.redisUrl.toString() }],
-        }),
-        inject: [ConfigService],
-      },
-      true, // isGlobal
-    ),
-    BullModule.forRootAsync({
-      imports: [ConfigModule],
-      useFactory: (configService: ConfigService) => {
-        // Note: BullMQ doesn't honor a URL for the Redis connection, and
-        // JS URL doesn't parse 'redis://' as a valid protocol, so we fool
-        // it by changing the URL to use 'http://' in order to parse out
-        // the host, port, username, password, etc.
-        // We could pass REDIS_HOST, REDIS_PORT, etc, in the environment, but
-        // trying to keep the # of environment variables from proliferating
-        const url = new URL(configService.redisUrl.toString().replace(/^redis[s]*/, 'http'));
-        const { hostname, port, username, password, pathname } = url;
-        return {
-          connection: {
-            host: hostname || undefined,
-            port: port ? Number(port) : undefined,
-            username: username || undefined,
-            password: password || undefined,
-            db: pathname?.length > 1 ? Number(pathname.slice(1)) : undefined,
-          },
-        };
-      },
-      inject: [ConfigService],
-    }),
-    BullModule.registerQueue(
-      {
-        name: QueueConstants.GRAPH_CHANGE_NOTIFY_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: true,
-          removeOnFail: false,
-        },
-      },
-      {
-        name: QueueConstants.GRAPH_CHANGE_PUBLISH_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: true,
-          removeOnFail: false,
-        },
-      },
-    ),
   ],
   providers: [GraphUpdatePublisherService, NonceService],
-  exports: [BullModule, GraphUpdatePublisherService],
+  exports: [GraphUpdatePublisherService, NonceService],
 })
 export class GraphUpdatePublisherModule {}
