+++ services/graph/apps/worker/src/reconnection_processor/graph.reconnection.processor.module.ts
@@ -1,68 +1,12 @@
-import { BullModule } from '@nestjs/bullmq';
 import { Module } from '@nestjs/common';
-import { RedisModule } from '@songkeys/nestjs-redis';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-import * as QueueConstants from '#lib/utils/queues';
-import { ConfigModule, ConfigService } from '#lib';
 import { GraphReconnectionService } from './graph.reconnection.processor.service';
 
 @Module({
   imports: [
-    ConfigModule,
     EventEmitterModule,
-    RedisModule.forRootAsync(
-      {
-        imports: [ConfigModule],
-        useFactory: (configService: ConfigService) => ({
-          config: [{ url: configService.redisUrl.toString() }],
-        }),
-        inject: [ConfigService],
-      },
-      true, // isGlobal
-    ),
-    BullModule.forRootAsync({
-      imports: [ConfigModule],
-      useFactory: (configService: ConfigService) => {
-        // Note: BullMQ doesn't honor a URL for the Redis connection, and
-        // JS URL doesn't parse 'redis://' as a valid protocol, so we fool
-        // it by changing the URL to use 'http://' in order to parse out
-        // the host, port, username, password, etc.
-        // We could pass REDIS_HOST, REDIS_PORT, etc, in the environment, but
-        // trying to keep the # of environment variables from proliferating
-        const url = new URL(configService.redisUrl.toString().replace(/^redis[s]*/, 'http'));
-        const { hostname, port, username, password, pathname } = url;
-        return {
-          connection: {
-            host: hostname || undefined,
-            port: port ? Number(port) : undefined,
-            username: username || undefined,
-            password: password || undefined,
-            db: pathname?.length > 1 ? Number(pathname.slice(1)) : undefined,
-          },
-        };
-      },
-      inject: [ConfigService],
-    }),
-    BullModule.registerQueue(
-      {
-        name: QueueConstants.RECONNECT_REQUEST_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: false,
-          removeOnFail: false,
-          attempts: 3,
-        },
-      },
-      {
-        name: QueueConstants.GRAPH_CHANGE_REQUEST_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: false,
-          removeOnFail: false,
-          attempts: 3,
-        },
-      },
-    ),
   ],
   providers: [GraphReconnectionService],
-  exports: [BullModule, GraphReconnectionService],
+  exports: [GraphReconnectionService],
 })
 export class GraphReconnectionModule {}
