+++ services/graph/apps/worker/src/request_processor/request.processor.module.ts
@@ -1,79 +1,12 @@
-/*
-https://docs.nestjs.com/modules
-*/
-
-import { BullModule } from '@nestjs/bullmq';
 import { Module } from '@nestjs/common';
-import { RedisModule } from '@songkeys/nestjs-redis';
-import * as QueueConstants from '#lib/utils/queues';
-import { BlockchainModule, ConfigModule, ConfigService, GraphStateManager } from '#lib';
+import { BlockchainModule, GraphStateManager } from '#lib';
 import { RequestProcessorService } from './request.processor.service';
 
 @Module({
   imports: [
-    ConfigModule,
     BlockchainModule,
-    RedisModule.forRootAsync(
-      {
-        imports: [ConfigModule],
-        useFactory: (configService: ConfigService) => ({
-          config: [{ url: configService.redisUrl.toString() }],
-        }),
-        inject: [ConfigService],
-      },
-      true, // isGlobal
-    ),
-    BullModule.forRootAsync({
-      imports: [ConfigModule],
-      useFactory: (configService: ConfigService) => {
-        // Note: BullMQ doesn't honor a URL for the Redis connection, and
-        // JS URL doesn't parse 'redis://' as a valid protocol, so we fool
-        // it by changing the URL to use 'http://' in order to parse out
-        // the host, port, username, password, etc.
-        // We could pass REDIS_HOST, REDIS_PORT, etc, in the environment, but
-        // trying to keep the # of environment variables from proliferating
-        const url = new URL(configService.redisUrl.toString().replace(/^redis[s]*/, 'http'));
-        const { hostname, port, username, password, pathname } = url;
-        return {
-          connection: {
-            host: hostname || undefined,
-            port: port ? Number(port) : undefined,
-            username: username || undefined,
-            password: password || undefined,
-            db: pathname?.length > 1 ? Number(pathname.slice(1)) : undefined,
-          },
-        };
-      },
-      inject: [ConfigService],
-    }),
-    BullModule.registerQueue(
-      {
-        name: QueueConstants.GRAPH_CHANGE_REQUEST_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: false,
-          removeOnFail: false,
-          attempts: 3,
-        },
-      },
-      {
-        name: QueueConstants.GRAPH_CHANGE_PUBLISH_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: true,
-          removeOnFail: false,
-          attempts: 1,
-        },
-      },
-      {
-        name: QueueConstants.RECONNECT_REQUEST_QUEUE,
-        defaultJobOptions: {
-          removeOnComplete: false,
-          removeOnFail: false,
-          attempts: 3,
-        },
-      },
-    ),
   ],
   providers: [RequestProcessorService, GraphStateManager],
-  exports: [BullModule, RequestProcessorService],
+  exports: [RequestProcessorService],
 })
 export class RequestProcessorModule {}
