+++ services/graph/apps/worker/src/request_processor/request.processor.service.ts
@@ -9,7 +9,7 @@ import { Option, Vec } from '@polkadot/types';
 import { AnyNumber } from '@polkadot/types/types';
 import { MILLISECONDS_PER_SECOND } from 'time-constants';
 import { BaseConsumer } from '../BaseConsumer';
-import * as QueueConstants from '#lib/utils/queues';
+import * as QueueConstants from '#lib/queues/queue-constants';
 import fs from 'fs';
 import {
   GraphStateManager,
