+++ services/graph/env.template
@@ -67,3 +67,6 @@ HEALTH_CHECK_MAX_RETRY_INTERVAL_SECONDS=10
 
 # Number of `/health` endpoint failures allowed before marking the provider webhook service down
 HEALTH_CHECK_MAX_RETRIES=4
+
+# Prefix to use for Redis cache keys
+CACHE_KEY_PREFIX=graph-service:
