+++ services/graph/libs/common/src/config/config.module.ts
@@ -1,8 +1,9 @@
 import { ConfigModule as NestConfigModule } from '@nestjs/config';
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
 import { ConfigService } from './config.service';
 import { configModuleOptions } from './env.config';
 
+@Global()
 @Module({
   imports: [NestConfigModule.forRoot(configModuleOptions)],
   controllers: [],
