+++ services/graph/libs/common/src/queues/queue.module.ts
@@ -1,0 +1,75 @@
+import { BullModule } from '@nestjs/bullmq';
+import { Global, Module } from '@nestjs/common';
+import { ConfigService } from '#lib/config/config.service';
+import * as QueueConstants from './queue-constants';
+import { SECONDS_PER_BLOCK } from '#lib/constants';
+import { MILLISECONDS_PER_SECOND } from 'time-constants';
+
+@Global()
+@Module({
+  imports: [
+    BullModule.forRootAsync({
+      useFactory: (configService: ConfigService) => {
+        // Note: BullMQ doesn't honor a URL for the Redis connection, and
+        // JS URL doesn't parse 'redis://' as a valid protocol, so we fool
+        // it by changing the URL to use 'http://' in order to parse out
+        // the host, port, username, password, etc.
+        // We could pass REDIS_HOST, REDIS_PORT, etc, in the environment, but
+        // trying to keep the # of environment variables from proliferating
+        const url = new URL(configService.redisUrl.toString().replace(/^redis[s]*/, 'http'));
+        const { hostname, port, username, password, pathname } = url;
+        return {
+          connection: {
+            host: hostname || undefined,
+            port: port ? Number(port) : undefined,
+            username: username || undefined,
+            password: password || undefined,
+            db: pathname?.length > 1 ? Number(pathname.slice(1)) : undefined,
+          },
+          prefix: `${configService.cacheKeyPrefix}:bull`,
+        };
+      },
+      inject: [ConfigService],
+    }),
+    BullModule.registerQueue(
+      {
+        name: QueueConstants.GRAPH_CHANGE_REQUEST_QUEUE,
+        defaultJobOptions: {
+          removeOnComplete: true,
+          removeOnFail: false,
+          attempts: 3,
+        },
+      },
+      {
+        name: QueueConstants.GRAPH_CHANGE_PUBLISH_QUEUE,
+        defaultJobOptions: {
+          removeOnComplete: true,
+          removeOnFail: false,
+          attempts: 1,
+        },
+      },
+      {
+        name: QueueConstants.GRAPH_CHANGE_NOTIFY_QUEUE,
+        defaultJobOptions: {
+          removeOnComplete: true,
+          removeOnFail: false,
+          attempts: 10,
+          delay: SECONDS_PER_BLOCK * MILLISECONDS_PER_SECOND,
+          backoff: {
+            type: 'fixed',
+            delay: SECONDS_PER_BLOCK * MILLISECONDS_PER_SECOND,
+          },
+        },
+      },
+      {
+        name: QueueConstants.RECONNECT_REQUEST_QUEUE,
+        defaultJobOptions: {
+          removeOnComplete: false,
+          removeOnFail: false,
+          attempts: 3,
+        },
+      },
+    )],
+  exports: [BullModule],
+})
+export class QueueModule {}
