+++ services/graph/libs/common/src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './base-consumer';
 export * from './processing';
-export * from './queues';
+export * from '../queues/queue-constants';
 export * from './redis';
