+++ services/graph/package.json
@@ -141,6 +141,7 @@
       "^#lib/config(|/.*)$": "<rootDir>/libs/common/src/config/$1",
       "^#lib/dtos(|/.*)$": "<rootDir>/libs/common/src/dtos/$1",
       "^#lib/interfaces(|/.*)$": "<rootDir>/libs/common/src/interfaces/$1",
+      "^#lib/queues(|/.*)$": "<rootDir>/libs/common/src/queues/$1",
       "^#lib/services(|/.*)$": "<rootDir>/libs/common/src/services/$1",
       "^#lib/types(|/.*)$": "<rootDir>/libs/common/src/types/$1",
       "^#lib/utils(|/.*)$": "<rootDir>/libs/common/src/utils/$1"
